requirements.txt
# Core Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0

# AI Provider SDKs
openai==1.3.5
anthropic==0.7.1
google-generativeai==0.3.1
groq==0.4.1

# Local Model Inference
llama-cpp-python==0.2.20  # For GGUF models
transformers==4.35.2
torch==2.1.1
accelerate==0.25.0

# NLP & Evaluation
nltk==3.8.1
textblob==0.17.1
vaderSentiment==3.3.2
pronouncing==0.2.0  # For rhyme detection
syllables==1.0.8

# Database & Data Processing
sqlalchemy==2.0.23
pandas==2.1.3
numpy==1.26.2

# Utilities
requests==2.31.0
aiohttp==3.9.1
python-multipart==0.0.6
pyyaml==6.0.1

# Development
pytest==7.4.3
black==23.11.0
flake8==6.1.0

# Optional: Music Production Integration
mido==1.3.0  # MIDI file processing
music21==9.1.0  # Music theory library
